% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FunctionalClust.Rmd.
\name{cv_functional_clust}
\alias{cv_functional_clust}
\title{Cross validation to choose the best h_0}
\usage{
cv_functional_clust(
  data,
  label,
  grid = seq(0.1, 2.5, 0.1),
  num_clust,
  max_iter = 1000,
  min_gap = 0.1,
  nrep = 10
)
}
\arguments{
\item{data}{Data in matrix form(n * p). Can be the data result from prepare_data}

\item{label}{Label set. Can be the label result from prepare_data}

\item{grid}{Grid values of bandwidth h}

\item{num_clust}{Number of cluster. If not specified, will equal to number of distinct label}

\item{max_iter}{Maximum iteration}

\item{min_gap}{Minimum difference of likelihood}

\item{nrep}{Number of repetitions}

\item{bandwidth}{Smooth bandwidth}
}
\value{
A list with three elements:
• all_lost – a list contains all the lost at all grid
• average_lost – average lost at all grid
• best_h – the best bandwidth
}
\description{
Cross validation to choose the best h_0
}