% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FunctionalClust.Rmd.
\name{functional_cluster}
\alias{functional_cluster}
\title{Main function of semi-supervised functional clustering}
\usage{
functional_cluster(
  data,
  label = NULL,
  num_clust,
  bandwidth = 1,
  max_iter = 1000,
  min_gap = 1,
  nrep = 5
)
}
\arguments{
\item{data}{Data in matrix form(n * p). Can be the data result from prepare_data}

\item{label}{Label set. Can be the label result from prepare_data}

\item{num_clust}{Number of cluster. If not specified, will equal to number of distinct label}

\item{bandwidth}{Smooth bandwidth}

\item{max_iter}{Maximum iteration}

\item{min_gap}{Minimum difference of likelihood}

\item{nrep}{Number of repetitions}
}
\value{
A list with three elements:
• data   – an n-column numeric matrix
• labels – a data.frame with
- index  : row numbers of the non-NA labels
- label  : integer codes (factor levels) of those labels
}
\description{
Main function of semi-supervised functional clustering
}