% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FunctionalClust.Rmd.
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Row normalized a matrix with sum up to one}
\usage{
gaussian_kernel(x, x_prime, sigma = bandwidth)
}
\description{
Row normalized a matrix with sum up to one
}