% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FunctionalClust.Rmd.
\name{prediction}
\alias{prediction}
\title{Make prediction on result$resp}
\usage{
prediction(res)
}
\arguments{
\item{resp}{The n*K memebership matrix generate from functional_cluster. It is in res$result$resp}
}
\value{
A dataframe with the first column is predictive label and the second colunm is proability
}
\description{
Make prediction on result$resp
}